requirements.txt
pymongo>=4.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
gspread>=5.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
